Rename ToolCallStrategyResolver to ToolCallBatcher
