+++ src/parlant/adapters/nlp/openai_service.py
@@ -35,7 +35,7 @@ import tiktoken
 
 from parlant.adapters.nlp.common import normalize_json_output
 from parlant.core.engines.alpha.prompt_builder import PromptBuilder
-from parlant.core.engines.alpha.tool_caller import SingleStrategyToolCallInferenceSchema
+from parlant.core.engines.alpha.tool_calling.single_tool_batch import SingleToolBatchSchema
 from parlant.core.loggers import Logger
 from parlant.core.nlp.policies import policy, retry
 from parlant.core.nlp.tokenization import EstimatingTokenizer
@@ -410,7 +410,7 @@ class OpenAIService(NLPService):
 
     @override
     async def get_schematic_generator(self, t: type[T]) -> OpenAISchematicGenerator[T]:
-        if t == SingleStrategyToolCallInferenceSchema:
+        if t == SingleToolBatchSchema:
             return GPT_4o[t](self._logger)  # type: ignore
         return GPT_4o_24_08_06[t](self._logger)  # type: ignore
